# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for CompStats - stat tracking for online melee game tournaments.",
        "title": "CompStats API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/admin/halfs": {
            "get": {
                "description": "get all halfs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Halfs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Half"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "create a half",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create half",
                "parameters": [
                    {
                        "description": "create json",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/half.createDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/database.Half"
                        }
                    }
                }
            }
        },
        "/admin/halfs/{id}": {
            "get": {
                "description": "get half by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a half",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Half ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Half"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete half by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a half",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Half ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "patch": {
                "description": "update half by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a half",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Half ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update json",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/half.updateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Half"
                        }
                    }
                }
            }
        },
...
        "/auth/login": {
            "post": {
                "description": "user login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "login json",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.loginDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.User"
                        }
                    }
                }
            }
        },
...
        "/auth/register": {
            "post": {
                "description": "user registration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "description": "register json",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.registerDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/database.User"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.loginDTO": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "auth.registerDTO": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "database.Half": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mapName": {
                    "type": "string"
                },
                "team1": {
                    "type": "string"
                },
                "team2": {
                    "type": "string"
                },
                "tournamentId": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "database.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "half.createDTO": {
            "type": "object",
            "required": [
                "gameType",
                "mapName",
                "team1",
                "team2",
                "tournamentId"
            ],
            "properties": {
                "gameType": {
                    "type": "string"
                },
                "mapName": {
                    "type": "string"
                },
                "team1": {
                    "type": "string"
                },
                "team2": {
                    "type": "string"
                },
                "tournamentId": {
                    "type": "integer"
                }
            }
        },
        "half.updateDTO": {
            "type": "object",
            "properties": {
                "mapName": {
                    "type": "string"
                },
                "team1": {
                    "type": "string"
                },
                "team2": {
                    "type": "string"
                }
            }
        },
...
    }
}